Refactoring: Moved NgModule class to model folder. Now detecting is module is a routing module.
